PASS NOTIFICATION INTO VERIFY WEBHOOK SIGNATURE IN PAYPAL PLUGIN
The process flow in sites that use the function to verify a webhook
signature typically requires the notification for subsequent process
steps. So, I'm changing to derive the notification in that process flow
in the site and pass it through to this function rather than deriving it
within this function.