REMOVE ABILITY TO PREFIX MESSAGES BUT ADD THE ABILITY TO OUTPUT DIVIDERS
The prefix capability is not needed now that log messages are output to
a separate file rather than mixed in with the system log messages. That
also makes the flow of the process being logged clearer and the ability
to put in dividers makes it possible to mark the beginning and end of
process flows.